+++ src/detection/board/board_nosupport.c
@@ -1,6 +1,6 @@
 #include "board.h"
 
-const char* ffDetectBoard(FFBoardResult* board)
+const char* ffDetectBoard(FF_MAYBE_UNUSED FFBoardResult* board)
 {
     return "Not supported on this platform";
 }