+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changes:
 * Drop the dependency of cJSON. We now use [yyjson](https://ibireme.github.io/yyjson/doc/doxygen/html/index.html) to parse JSON documents.
 * Remove `--shell-version` and `--terminal-version`. They are always enabled
 * Remove `--*-error-format`
+* Remove `--display-detect-name`. Display name is always detected, and will be printed if multiple displays are detected
 
 Features:
 * FreeBSD support is improved greatly, and actually tested in a phycial machine
@@ -14,7 +15,7 @@ Features:
 * Support MATE Terminal version & terminal font detection
 * Add `--no-buffer` option for easier debugging. CMake option `ENABLE_BUFFER` is removed and always enabled.
 * Support `--*-key-color` option to change the key color of specified module
-* Support `--colors-symbol`
+* Support `--colors-symbol` and `--colors-padding-left`
 
 # 1.11.3